+++ CHANGELOG.rst
@@ -2,8 +2,9 @@
 0.9.8 (unreleased)
 ------------------
 
-- Add the `pytest.mark.target` marker in the place of `pytest_nodev.search`.
+- Add the ``pytest.mark.target`` marker in the place of ``pytest_nodev.search``.
   Issue `#28 <https://github.com/nodev-io/pytest-nodev/issues/28>`_.
+- Rename main fixture to ``candidate`` from ``wish``.
 
 
 0.9.7 (2016-03-13)